Refactored to streamline the conversion process and make it easier to explain.
